Int Tensor

+++ src/pass/finalization/sparsity.cpp
@@ -141,7 +141,7 @@ void setSparsityFnc(const mv::pass::PassEntry& pass, mv::ComputationModel& model
 
             //Create Tensor with Sparsity Data
             mv::Shape sparsityShape(ndims);
-            std::vector<double> data(sparsityShape.totalSize(), 0);
+            std::vector<int64_t> data(sparsityShape.totalSize(), 0);
             //mv::Tensor sparsityTensor("backup", sparsityShape, mv::DType("UInt8"), mv::Order("WHCN"), data);
             auto sparsityTensor = dm.defineTensor(opIterator->getName() + "_sparse_dw", sparsityShape, mv::DType("UInt8"), mv::Order("NCHW"), data);
 
@@ -149,7 +149,7 @@ void setSparsityFnc(const mv::pass::PassEntry& pass, mv::ComputationModel& model
                 for(unsigned ky = 0; ky < sparsityShape[1]; ++ky)
                     for(unsigned ic = 0; ic < sparsityShape[2]; ++ic)
                         for(unsigned oc = 0; oc < sparsityShape[3]; ++oc)
-                            sparsityTensor->at({kx, ky, ic, oc}) = perChannelSparsity[ky*sparsityShape[0] + kx];
+                            sparsityTensor->at({kx, ky, ic, oc}) = static_cast<int64_t>(perChannelSparsity[ky*sparsityShape[0] + kx]);
 
             //Add sparsity map to conv
             om.addAttr(opIterator, "sparsityMap", sparsityTensor->getName());
